tests/wpt/web-platform-tests/navigation-timing/nav2_test_instance_accessible_from_the_start.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var navigationTiming2 = window.performance.getEntriesByType('navigation')[0];
  assert_true(navigationTiming2 !== undefined);
  assert_equals(navigationTiming2.entryType, "navigation");
  assert_equals(navigationTiming2.startTime, 0);
  assert_equals(navigationTiming2.initiatorType, "navigation");
  done();
}, "PerformanceNavigationTiming instance exists with reasonable values.");
</script>
